:truck: Restructure files
