+++ src/Unsplash/Api/Collections/RequestModels/GetCollectionPhotosParams.cs
