+++ src/Unsplash/Api/Collections/RequestModels/ListCollectionsParams.cs
