+++ src/Unsplash/Api/Photos/RequestModels/ContentFilter.cs
@@ -1,0 +1,12 @@
+﻿using System.Runtime.Serialization;
+
+namespace Unsplash.Api.Photos
+{
+    public enum ContentFilter
+    {
+        [EnumMember(Value = "low")]
+        Low,
+        [EnumMember(Value = "high")]
+        High
+    }
+}
