+++ src/Unsplash/Api/Photos/RequestModels/Orientation.cs
@@ -1,0 +1,14 @@
+﻿using System.Runtime.Serialization;
+
+namespace Unsplash.Api.Photos
+{
+    public enum Orientation
+    {
+        [EnumMember(Value = "landscape")]
+        Landscape,
+        [EnumMember(Value = "portrait")]
+        Portrait,
+        [EnumMember(Value = "squarish")]
+        Squarish
+    }
+}
