+++ src/Unsplash/Api/Search/RequestModels/SearchCollectionsParams.cs
