+++ src/Unsplash/Api/Stats/IStatsApi.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+using Unsplash.Models;
+
+namespace Unsplash.Api.Stats
+{
+    public interface IStatsApi
+    {
+        Task<StatsTotals> GetTotalsAsync();
+        Task<PastMonthStats> GetPastMonthStatsAsync();
+    }
+}
