+++ src/Unsplash/Api/Users/RequestModels/GetUserPhotosParams.cs
