+++ src/Unsplash/Models/PaginatedList.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace Unsplash.Api.Search
+namespace Unsplash.Models
 {
     public class PaginatedList<T>
     {
